streamai/requirements.txt
asgiref==3.11.1
Django==5.2.11
djangorestframework==3.16.1
gunicorn==25.0.1
packaging==26.0
pillow==12.1.0
sqlparse==0.5.5
